evaldb/requirements.txt
attrs==21.2.0
importlib-metadata==4.8.1
importlib-resources==5.2.2
iniconfig==1.1.1
packaging==21.0
pluggy==1.0.0
py==1.10.0
pyInovEval==0.1.0
pyodbc==4.0.32
pyparsing==2.4.7
pytest==6.2.5
PyYAML==5.4.1
toml==0.10.2
typing-extensions==3.10.0.2
zipp==3.5.0
